Fix comments in queries causing bad results
ColumnStore modifies the query string stored in the thd object so
that it can be used by ColumnStore's parser. The biggest operation
here is to remove newline characters. This doesn't remove inline
comments before removing the newline characters so an inline
comment turns the rest of the query into a comment when the
newline characters are dropped.

The effect of this can be seen when ColumnStore creates an aria
temporary table, it builds a query to do this using the processed
version of the current query.

This commit removes the SQL comments during the query text
processing.