MCOL-424 Disable indexes for cross-engine
Index lookups are not part of the pushed down condition so do not turn
into 'where' conditions for cross-engine joins. This fix disables
indexes for tables where there is a ColumnStore table in use so the
condition is pushed down properly.

When the cross-engine execution happens it will be a basic query
against a standard storage engine so indexes are used.