MCOL-441 Fix segfault on SP error
When a stored procedure errors it can cause index_hints to be used after
thd->mem_root is freed triggering a segfault. This is a regression
caused by the fix of MCOL-424.

index_hints are now created in less cases and set to NULL upon error
(will be freed later when mem_root is freed).