MCOL-618 Fix prepared statements
The table pointer in the table list can point to junk if the table is
derived, an I_S table or a view. For prepared statements we relied on
this information being correct.

With this patch derived and I_S are ignored since they will never be
ColumnStore tables. In addition views are processed to check their
underlying tables for ColumnStore tables.