MCOL-650 Prep stmt + CALL use vtable
With SQLCOM_CALL (stored procedures) prepared statements were falling
through to vtable disable straight away. This changes the behaviour to
the same as normal queries, that is to inspect the first query of a
stored procedure and execute using vtable if it is ColumnStore.

The change rearranges sql_prepare as we don't want to fall into
thd->get_command() == COM_STMT_EXECUTE before checking for SQLCOM_CALL.